Address feedback; move FileDownloader.cs to /Shared
