+++ src/Components/Server/src/DependencyInjection/ComponentServiceCollectionExtensions.cs
@@ -68,6 +68,7 @@ public static class ComponentServiceCollectionExtensions
         services.TryAddSingleton<CircuitIdFactory>();
         services.TryAddScoped<IErrorBoundaryLogger, RemoteErrorBoundaryLogger>();
         services.TryAddScoped<IFileDownloader, FileDownloader>();
+
         services.TryAddScoped(s => s.GetRequiredService<ICircuitAccessor>().Circuit);
         services.TryAddScoped<ICircuitAccessor, DefaultCircuitAccessor>();
